+++ src/Tests/Dfe.Complete.Api.Tests.Integration/Controllers/TasksDataController/TaskDataControllerTests.cs
@@ -921,40 +921,6 @@ namespace Dfe.Complete.Api.Tests.Integration.Controllers.TasksDataController
         }
         [Theory]
         [CustomAutoData(typeof(CustomWebApplicationDbContextFactoryCustomization))]
-        public async Task UpdateLandRegistryTitlePlansTaskAsync_ShouldUpdate_ConversionTaskData(
-            CustomWebApplicationDbContextFactory<Program> factory,
-            ITasksDataClient tasksDataClient,
-            UpdateLandRegistryTitlePlansTaskCommand command,
-            IFixture fixture)
-        {
-            // Arrange
-            factory.TestClaims = [new Claim(ClaimTypes.Role, ApiRoles.ReadRole), new Claim(ClaimTypes.Role, ApiRoles.UpdateRole), new Claim(ClaimTypes.Role, ApiRoles.WriteRole)];
-
-            var dbContext = factory.GetDbContext<CompleteContext>();
-
-            var taskData = fixture.Create<ConversionTasksData>();
-            dbContext.ConversionTasksData.Add(taskData);
-
-            await dbContext.SaveChangesAsync();
-            command.TaskDataId = new TaskDataId { Value = taskData.Id.Value };
-            command.Received = true;
-            command.Cleared = true;
-            command.Saved = false;
-
-            // Act
-            await tasksDataClient.UpdateLandRegistryTitlePlansTaskAsync(command, default);
-
-            // Assert
-            dbContext.ChangeTracker.Clear();
-            var existingTaskData = await dbContext.ConversionTasksData.SingleOrDefaultAsync(x => x.Id == taskData.Id);
-            Assert.NotNull(existingTaskData);
-            Assert.True(existingTaskData.LandRegistryReceived);
-            Assert.True(existingTaskData.LandRegistryCleared);
-            Assert.False(existingTaskData.LandRegistrySaved);
-        }
-
-        [Theory]
-        [CustomAutoData(typeof(CustomWebApplicationDbContextFactoryCustomization))]
         public async Task UpdateCommercialTransferAgreementTaskAsync_ShouldUpdate_TransferTaskData(
           CustomWebApplicationDbContextFactory<Program> factory,
           ITasksDataClient tasksDataClient,
@@ -1049,10 +1015,13 @@ namespace Dfe.Complete.Api.Tests.Integration.Controllers.TasksDataController
 
             await dbContext.SaveChangesAsync();
             command.TaskDataId = new TaskDataId { Value = taskData.Id.Value };
-            command.Received = true;
-            command.Signed = true;
-            command.Saved = false;
-            command.Cleared = false;
+            command.LandQuestionnaireReceived = true;
+            command.LandQuestionnaireSigned = true;
+            command.LandQuestionnaireSaved = false;
+            command.LandQuestionnaireCleared = false;
+            command.LandRegistryTitlePlansReceived = true;
+            command.LandRegistryTitlePlansCleared = true;
+            command.LandRegistryTitlePlansSaved = false;
 
             // Act
             await tasksDataClient.UpdateLandQuestionnaireTaskAsync(command, default);
@@ -1065,6 +1034,9 @@ namespace Dfe.Complete.Api.Tests.Integration.Controllers.TasksDataController
             Assert.True(existingTaskData.LandQuestionnaireReceived);
             Assert.True(existingTaskData.LandQuestionnaireSigned);
             Assert.False(existingTaskData.LandQuestionnaireSaved);
+            Assert.True(existingTaskData.LandRegistryReceived);
+            Assert.True(existingTaskData.LandRegistryCleared);
+            Assert.False(existingTaskData.LandRegistrySaved);
         }